% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{append_nav_to_bt_results}
\alias{append_nav_to_bt_results}
\title{Append NAV to backtest results dataframe}
\usage{
append_nav_to_bt_results(bt_results_df)
}
\arguments{
\item{bt_results_df}{Results dataframe from rsims backtest}
}
\value{
New results data frame with a row for portfolio NAV each day
}
\description{
Append NAV to backtest results dataframe
}
\examples{
\dontrun{
results <- cash_backtest(prices, weights)
results <- append_nav_to_bt_results(results)
}
}
