% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{calc_ann_turnover}
\alias{calc_ann_turnover}
\title{Annual turnover}
\usage{
calc_ann_turnover(results_df, mean_equity, start_date, end_date)
}
\arguments{
\item{results_df}{Results dataframe from rsims backtest}

\item{mean_equity}{Mean daily portfolio equity}

\item{start_date}{Backtest start date}

\item{end_date}{Backtest end date}
}
\value{
Annual turnover as yearly average of total sell trades as pct of mean equity
}
\description{
Annual turnover
}
\examples{
\dontrun{
results_df <- cash_backtest(prices, weights)
port_returns <- results_df \%>\%
  calc_port_returns()
mean_equity <- mean(port_returns$totalequity)
results_df \%>\%
  calc_ann_turnover(mean_equity, "2000-01-01", "2021-12-31")
}
}
