% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{calc_port_returns}
\alias{calc_port_returns}
\title{Timeseries portfolio returns}
\usage{
calc_port_returns(results_df)
}
\arguments{
\item{results_df}{Results dataframe from rsims backtest}
}
\value{
Time series of portfolio returns
}
\description{
Timeseries portfolio returns
}
\examples{
\dontrun{
cash_backtest(prices, weights) \%>\%
  calc_port_returns()
}
}
