% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{calc_sharpe}
\alias{calc_sharpe}
\title{Sharpe ratio}
\usage{
calc_sharpe(results_df)
}
\arguments{
\item{results_df}{Results dataframe from rsims backtest}
}
\value{
Annualised Sharpe ratio
}
\description{
Sharpe ratio
}
\examples{
\dontrun{
cash_backtest(prices, weights) \%>\%
  calc_sharpe()
}
}
