% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_commission_backtest.R
\name{cash_backtest}
\alias{cash_backtest}
\title{For backwards compatibility
Follows renaming of original cash_backtest function to fixed_commission_backtest}
\usage{
cash_backtest(
  prices,
  theo_weights,
  trade_buffer = 0,
  initial_cash = 10000,
  commission_pct = 0,
  capitalise_profits = FALSE
)
}
\description{
For backwards compatibility
Follows renaming of original cash_backtest function to fixed_commission_backtest
}
