% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{combine_port_asset_returns}
\alias{combine_port_asset_returns}
\title{Combine portfolio and asset returns}
\usage{
combine_port_asset_returns(results_df, returns_df)
}
\arguments{
\item{results_df}{Results dataframe from rsims backtest}

\item{returns_df}{Dataframe of asset returns}
}
\value{
Combined dataframe of portfolio and asset returns
}
\description{
Combine portfolio and asset returns
}
