% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futs_returns_series.R
\name{days_to_expiry}
\alias{days_to_expiry}
\title{Number of Days to Expiry}
\usage{
days_to_expiry(contracts)
}
\arguments{
\item{contracts}{dataframe of futures contract with at least
ticker/date fields. date must be a date object.}
}
\value{
dataframe of futures contracts with date/dte fields
}
\description{
Add dte column to dataframe of futures contract prices. Here,
dte refers to calendar days to expiry
}
