% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_commission_backtest.R
\name{fixed_commission_backtest}
\alias{fixed_commission_backtest}
\title{Fixed Commission Backtest}
\usage{
fixed_commission_backtest(
  prices,
  theo_weights,
  trade_buffer = 0,
  initial_cash = 10000,
  commission_pct = 0,
  capitalise_profits = FALSE
)
}
\arguments{
\item{prices}{Matrix of trade prices. Column 1 must be the timestamp or index.}

\item{theo_weights}{Matrix of theoretical weights. Column 1 must be the timestamp or index.}

\item{trade_buffer}{Trade buffer parameter}

\item{initial_cash}{Inital cash balance}

\item{commission_pct}{Percent commission charged on trades}

\item{capitalise_profits}{If TRUE, utilise profits and initial cash balance in determining position sizes. If FALSE, profits accrue as a cash balance and are not reinvested.}
}
\value{
long dataframe of results - dates, trades, commissions, value of portfolio components
}
\description{
Fixed Commission Backtest
}
\details{
\code{theo_weights} should be date-aligned with \code{prices} - it is up to the user to lag \code{theo_weights} as necessary to
ensure that trades occur at appropriate prices.
}
