% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission_models.R
\name{fixed_percent}
\alias{fixed_percent}
\title{Fixed percent commission}
\usage{
fixed_percent(trades, trade_value, commission_pct = 0)
}
\arguments{
\item{trade_value}{vector of traded value for each asset}

\item{commission_pct}{percentage of traded value to be paid as commission (scalar or vector of same length as \code{trade_value})}
}
\value{
vector of commissions
}
\description{
Commission calculated as a fixed percent of traded value. Suitable for many
crypto exchanges.
}
\examples{
fixed_percent(c(150, 75, -220), 0.01)
}
