% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_commission_futs_backtest.R
\name{futs_backtest_equity}
\alias{futs_backtest_equity}
\title{Calculate timeseries of end-of-day portfolio equity from futures backtest results.}
\usage{
futs_backtest_equity(futs_backtest_results)
}
\arguments{
\item{futs_backtest_results}{dataframe output of
\code{fixed_commission_futs_backtest}}
}
\value{
dataframe of date, equity
}
\description{
End-of-day portfolio equity for a futures simulation is
equivalent to the cash balance (having settled cash from existing positions)
plus posted margin.
}
\examples{
/dontrun{futs_backtest_equity(results)}
}
