% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission_models.R
\name{futs_per_contract_commission}
\alias{futs_per_contract_commission}
\title{Per-Contract Futures Commissions}
\usage{
futs_per_contract_commission(contracts_traded, per_contract_commission)
}
\arguments{
\item{contracts_traded}{vector of number of contracts traded per symbol}

\item{per_contract_commission}{vector of per-contract commission per symbol}
}
\value{
vector of commissions paid per symbol
}
\description{
Calculates commission on number of contracts traded given a per-
contract commission structure.
}
\examples{
contracts_traded <- c("ES" = 5, "GC" = -2)
per_contract_commission <- c("ES" = 0.85, "GC" = 0.5)
futs_per_contract_commission(contracts_traded, per_contract_commission)
}
