% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{Generate ggplot colours}
\usage{
gg_colour_hue(n)
}
\arguments{
\item{n}{number of colours to generate}
}
\value{
Character vector of n default ggplot2 colours in hex format
}
\description{
Generate ggplot colours
}
\examples{
gg_colour_hue(10)
}
