% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_advection_diffusion_equation}
\alias{make_advection_diffusion_equation}
\title{Advection-Diffusion Equation: du/dt = -v\emph{du/dx + D}d2u/dx2}
\usage{
make_advection_diffusion_equation(v = 1, D = 1)
}
\description{
Advection-Diffusion Equation: du/dt = -v\emph{du/dx + D}d2u/dx2
}
