% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_advection_equation}
\alias{make_advection_equation}
\title{Pure Advection (Transport) Equation: du/dt = -v * du/dx}
\usage{
make_advection_equation(v = 1)
}
\description{
Pure Advection (Transport) Equation: du/dt = -v * du/dx
}
