% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_boundary_custom}
\alias{make_boundary_custom}
\title{Custom Boundary Conditions}
\usage{
make_boundary_custom(fun)
}
\arguments{
\item{fun}{function(u, t) returning the state after enforcing the
boundary condition.}
}
\description{
Custom Boundary Conditions
}
