% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_boundary_open}
\alias{make_boundary_open}
\title{Open (Outflow / Zero-Curvature) Boundary Conditions}
\usage{
make_boundary_open()
}
\description{
Open (Outflow / Zero-Curvature) Boundary Conditions
}
