% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_boundary_reflecting}
\alias{make_boundary_reflecting}
\title{Reflecting (Zero-Flux / Neumann) Boundary Conditions}
\usage{
make_boundary_reflecting()
}
\description{
Reflecting (Zero-Flux / Neumann) Boundary Conditions
}
