% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{make_col_scale}
\alias{make_col_scale}
\title{Portfolio color scale (tickers only)}
\usage{
make_col_scale(cols)
}
\arguments{
\item{cols}{Named character vector of colours in hex format. Names correspond
to tickers in the universe, plus "Cash" and "Portfolio"}
}
\value{
ggplot2 colour scale
}
\description{
Portfolio color scale (tickers only)
}
\examples{
my_cols <- portfolio_colours(7)
colour_scale <- make_col_scale(my_cols)
}
