% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_custom_equation}
\alias{make_custom_equation}
\title{Custom Equation}
\usage{
make_custom_equation(rhs, parameters = list(), name = "custom")
}
\arguments{
\item{rhs}{function(u, grid, t, parameters) returning du/dt.}
}
\description{
Custom Equation
}
