% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_diffusion_equation}
\alias{make_diffusion_equation}
\title{Diffusion (Heat) Equation: du/dt = D * d2u/dx2}
\usage{
make_diffusion_equation(D = 1)
}
\description{
Diffusion (Heat) Equation: du/dt = D * d2u/dx2
}
