% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_initial_gaussian}
\alias{make_initial_gaussian}
\title{Gaussian Initial Condition}
\usage{
make_initial_gaussian(grid, center, width, amplitude = 1, background = 0)
}
\description{
Gaussian Initial Condition
}
