% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_initial_point_source}
\alias{make_initial_point_source}
\title{Point-Source Initial Condition}
\usage{
make_initial_point_source(grid, location, amplitude = 100, background = 0)
}
\description{
Point-Source Initial Condition
}
