% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_initial_random}
\alias{make_initial_random}
\title{Random (White-Noise) Initial Condition}
\usage{
make_initial_random(grid, mean = 0, sd = 1)
}
\description{
Random (White-Noise) Initial Condition
}
