% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_splines.R
\name{make_radial_basis}
\alias{make_radial_basis}
\title{Create Gaussian Radial Basis Functions}
\usage{
make_radial_basis(x, knots = 8, length_scale = NULL)
}
\arguments{
\item{x}{Numeric predictor.}

\item{knots}{Number of radial basis centres or supplied knot values.}

\item{length_scale}{Width of Gaussian bumps.}
}
\value{
Matrix containing radial basis columns.
}
\description{
Creates localized Gaussian bump basis functions.
}
\details{
Each basis column is:

\deqn{
B_k(x)=exp(-1/2((x-\kappa_k)/l)^2)
}

where \eqn{\kappa_k} are knot locations.
}
