% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pdes.R
\name{make_reaction_diffusion_equation}
\alias{make_reaction_diffusion_equation}
\title{Reaction-Diffusion (Fisher-KPP / Logistic Growth) Equation:
du/dt = D\emph{d2u/dx2 + r}u*(1 - u/K)}
\usage{
make_reaction_diffusion_equation(D = 1, r = 1, K = 1)
}
\description{
Reaction-Diffusion (Fisher-KPP / Logistic Growth) Equation:
du/dt = D\emph{d2u/dx2 + r}u*(1 - u/K)
}
