% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_commission_futs_backtest.R
\name{make_sim_prices_matrix}
\alias{make_sim_prices_matrix}
\title{Make Futures Simulation Prices Matrix}
\usage{
make_sim_prices_matrix(wrangled_contracts)
}
\arguments{
\item{wrangled_contracts}{dataframe output of \code{wrangle_contracts_on_oi()}}
}
\value{
Wide numerical matrix with the following columns:
date: numerical days since epoch
close_current_contract_\code{*}: \code{n} columns designating the closing price
(close\emph{point_value) of the current contract for each symbol.
close_previous_contract_\code{*}: \code{n} columns designating the closing price
(close}point_value) of yesterday's contract for each symbol. Will have the
same value as close_current_contract unless there was a roll.
roll_\code{*}: \code{n} columns of (0,1) designating whether a roll occurs on the
current date (1) or not (0).
}
\description{
Transforms a wrangled contracts dataframe (the output of
\code{wrangle_contracts_on_oi}) into a wide numerical matrix for simulation.

Returns a matrix with 3 columns per symbol plus the date. Symbols appear as
suffixes in the column names (the \code{*} below).
}
\examples{
\dontrun{
wrangled <- wrangle_contracts_on_oi(futures)
sim_prices <- make_sim_prices_matrix(wrangled)
}
}
