% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_commission_futs_backtest.R
\name{make_sim_weights_matrix}
\alias{make_sim_weights_matrix}
\title{Make Futures Simulation Weights Matrix}
\usage{
make_sim_weights_matrix(weights)
}
\arguments{
\item{weights}{Long dataframe of symbol, date, target_weight}
}
\value{
Wide matrix of date, and columns for target weights for each symbol
}
\description{
Takes a long dataframe of symbol, date, target_weight and
returns a wide matrix of date, and columns for target weights for each symbol.

Helper function for wrangling data into the required shape for simulation.
}
\examples{
/dontrun{
weights <- data.frame(
 date = dates,
 symbol = symbols,
 target_weight = target_weights
)
make_sim_weights_matrix(weights)
}
}
