% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_commission_backtest.R
\name{min_commission_backtest}
\alias{min_commission_backtest}
\title{Equities Cash Accounting Backtest}
\usage{
min_commission_backtest(
  prices,
  unadjusted_prices,
  theo_weights,
  trade_buffer = 0,
  initial_cash = 10000,
  capitalise_profits = FALSE,
  rebal_tolerance,
  leverage = 1,
  commission_fun,
  ...
)
}
\arguments{
\item{prices}{Matrix of trade prices. Column 1 must be the timestamp or index.}

\item{theo_weights}{Matrix of theoretical weights. Column 1 must be the timestamp or index.}

\item{trade_buffer}{Trade buffer parameter}

\item{initial_cash}{Inital cash balance}

\item{capitalise_profits}{If TRUE, utilise profits and initial cash balance in determining position sizes. If FALSE, profits accrue as a cash balance and are not reinvested.}

\item{commission_pct}{Percent commission charged on trades}
}
\value{
long dataframe of results - dates, trades, commissions, value of portfolio components
}
\description{
Equities Cash Accounting Backtest
}
\details{
\code{theo_weights} should be date-aligned with \code{prices} - it is up to the user to lag \code{theo_weights} as necessary to
ensure that trades occur at appropriate prices.
}
