% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_commission_futs_backtest.R
\name{min_commission_futs_backtest}
\alias{min_commission_futs_backtest}
\title{Futures Backtest, roll on days to expiry, minimum commission model}
\usage{
min_commission_futs_backtest(
  prices,
  theo_weights,
  trade_buffer = 0,
  initial_cash = 10000,
  capitalise_profits = FALSE,
  leverage = 1,
  commission_fun,
  ...
)
}
\arguments{
\item{prices}{Matrix of trade prices. Column 1 must be the timestamp or
index. Column 2 must be days-to-expiry.}

\item{theo_weights}{Matrix of theoretical weights. Column 1 must be the
timestamp or index.}

\item{trade_buffer}{Trade buffer parameter}

\item{initial_cash}{Inital cash balance}

\item{capitalise_profits}{If TRUE, utilise profits and initial cash balance
in determining position sizes. If FALSE, profits accrue as a cash balance and
are not reinvested.}

\item{leverage}{Leverage to apply in simulation}

\item{commission_fun}{Function for determining commissions from prices and
trades}
}
\value{
long dataframe of results - dates, trades, commissions, value of portfolio components
}
\description{
Event-based simulation based on desired futures positions, raw
contract prices (ie not backadjusted) and user-specified roll date (dte).
}
\details{
\code{theo_weights} should be date-aligned with \code{prices} - it is up to the user to lag \code{theo_weights} as necessary to
ensure that trades occur at appropriate prices.
}
