% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{portfolio_colours}
\alias{portfolio_colours}
\title{Generate a set of portfolio colours}
\usage{
portfolio_colours(n)
}
\arguments{
\item{n}{number of tickers in portfolio}
}
\value{
Named character vector of n+2 default ggplot2 colours in hex format
corresponding to each ticker plus "Cash" and "Portfolio"
}
\description{
Generate a set of portfolio colours
}
\examples{
portfolio_colours(10)
}
