% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{positionsFromNoTradeBufferMinComm}
\alias{positionsFromNoTradeBufferMinComm}
\title{Calculate target positions from theoretical weights and trade buffer parameter in the presence of minimum commission}
\usage{
positionsFromNoTradeBufferMinComm(
  current_positions,
  current_prices,
  current_theo_weights,
  cap_equity,
  trade_buffer
)
}
\description{
Calculate target positions from theoretical weights and trade buffer parameter in the presence of minimum commission
}
