% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futs_returns_series.R
\name{roll_on_dte}
\alias{roll_on_dte}
\title{Roll on Days to Expiry}
\usage{
roll_on_dte(contracts, roll_dte = 1, roll_cost = 0)
}
\arguments{
\item{contracts}{List of dfs of futures with at least ticker/date/close columns}

\item{roll_dte}{Calendar days to expiry to roll into next contract}

\item{cost}{Percent cost of trading into next contract (total cost of
trading out of current contract and into next contract)}
}
\value{
A continuous futures return series
}
\description{
Create a continuous returns series from a set of futures
contracts by rolling on a given number of days to expiry.
}
