% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{rolling_ann_port_perf}
\alias{rolling_ann_port_perf}
\title{Rolling annualised portfolio performance}
\usage{
rolling_ann_port_perf(port_returns_df)
}
\arguments{
\item{port_returns_df}{Dataframe of portfolio returns}
}
\value{
Dataframe of rolling annualised portfolio statistics
}
\description{
Rolling annualised portfolio performance
}
