% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{rolling_ann_port_perf_plot}
\alias{rolling_ann_port_perf_plot}
\title{Rolling annualised portfolio performance plot}
\usage{
rolling_ann_port_perf_plot(perf_df)
}
\arguments{
\item{perf_df}{Dataframe of rolling annualised metrics}
}
\value{
ggplot2 of rolling annualised performance
}
\description{
Rolling annualised portfolio performance plot
}
\examples{
\dontrun{
results_df \%>\%
  calc_port_returns() \%>\%
  rolling_ann_port_perf() \%>\%
  rolling_ann_port_perf_plot()
}
}
