% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_latent.R
\name{simulate_correlated_latents}
\alias{simulate_correlated_latents}
\title{Simulate correlated latent variables}
\usage{
simulate_correlated_latents(
  n,
  Sigma = NULL,
  sds = NULL,
  Rho = NULL,
  means = NULL,
  names = NULL
)
}
\arguments{
\item{n}{Number of units (rows).}

\item{Sigma}{K x K covariance matrix. Either Sigma or (sds + Rho)
must be supplied.}

\item{sds}{Optional length-K vector of standard deviations, used
with Rho to build Sigma = diag(sds) \%\emph{\% Rho \%}\% diag(sds).}

\item{Rho}{Optional K x K correlation matrix, paired with sds.}

\item{means}{Length-K vector of latent means (default: zeros).}

\item{names}{Optional length-K character vector naming the latents
(default: eta1, eta2, ...).}
}
\value{
list(
data  = list(Eta = n x K matrix of latent draws),
truth = list(means = ..., Sigma = ..., Rho = ..., sds = ..., L = ..., K = ..., n = ...)
)
}
\description{
Generative process:
(eta_1, ..., eta_K)_i ~ MVN(mu, Sigma)     for i = 1..n
}
\details{
Internally uses the Cholesky factor L of Sigma (Sigma = L L^T) and
draws via the multivariate non-centered parameterization:
z_i    ~ Normal(0, I_K)
Eta_i   = mu + L \%*\% z_i
This is the multivariate analogue of simulate_latent()'s NCP and
avoids a hard dependency on MASS::mvrnorm.

Supply either \code{Sigma} directly, or \code{sds} + \code{Rho} (spread/wobble +
correlation), which is usually the more interpretable way to reason
about correlated constructs.
}
\examples{
Rho <- matrix(c(1, 0.6, 0.6, 1), 2, 2)
sim <- simulate_correlated_latents(n = 500, sds = c(1, 1), Rho = Rho,
                                    names = c("vigour", "stress_tolerance"))
}
