% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_latent.R
\name{simulate_latent}
\alias{simulate_latent}
\title{Simulate independent latent variable(s)}
\usage{
simulate_latent(n, mean = 0, sd = 1, ncp = TRUE)
}
\arguments{
\item{n}{Number of units (rows).}

\item{mean}{Mean mu of the latent distribution.}

\item{sd}{Standard deviation sigma of the latent distribution.}

\item{ncp}{Logical; TRUE (default) draws via the non-centered
parameterization eta_i = mu + sigma * z_i.}
}
\value{
list(
data  = list(eta = numeric vector, length n),
truth = list(mean = mean, sd = sd, n = n)
)
}
\description{
Generative process (Statistical Rethinking notation):
eta_i ~ Normal(mu, sigma)     for i = 1..n
}
\details{
Non-centered form (default):
z_i   ~ Normal(0, 1)
eta_i  = mu + sigma * z_i
}
\examples{
sim <- simulate_latent(n = 500, mean = 0, sd = 1)
str(sim$data$eta)
}
