% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{stacked_area_chart}
\alias{stacked_area_chart}
\title{Plot exposures and NAV timeseries as a stacked area chart}
\usage{
stacked_area_chart(results_df, title, tickers, colours)
}
\arguments{
\item{results_df}{Results dataframe from rsims backtest with daily NAV}

\item{title}{Plot title}

\item{tickers}{Character vector of tickers used in the simulation}

\item{colours}{Character vector of colours to use for tickers and Cash exposures}
}
\value{
A ggplot2 stacked area chart of timeseries of exposures and NAV
}
\description{
Plot exposures and NAV timeseries as a stacked area chart
}
\examples{
\dontrun{
tickers = unique(prices$ticker)
my_cols = portfolio_colours(length(tickers))
cash_backtest(prices, weights) \%>\%
  append_nav_to_bt_results() \%>\%
  stacked_area_chart("Exposures and NAV", tickers = tickers, colours = my_cols)
}
}
