% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{trades_chart}
\alias{trades_chart}
\title{Trade chart}
\usage{
trades_chart(results_df, title, fill_scale, colour_scale)
}
\arguments{
\item{results_df}{Results dataframe from rsims backtest}

\item{title}{Plot title}

\item{fill_scale}{ggplot2 fill scale (tickers only)}

\item{colour_scale}{ggplot2 colour scale (tickers only)}
}
\value{
Plot of trade value by ticker and date
}
\description{
Trade chart
}
\examples{
\dontrun{
tickers = unique(prices$ticker)
my_cols = portfolio_colours(length(tickers))
col_scale <- make_col_scale(my_cols)
fill_scale <- make_fill_scale(my_cols)
results <- cash_backtest(prices, weights)
trades_chart(results, "Trade Value", fill_scale, col_scale)
}
}
