% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission_models.R
\name{us_tiered_commission}
\alias{us_tiered_commission}
\title{US Tiered Minimum Commission Model}
\usage{
us_tiered_commission(
  shares_traded,
  prices,
  unadjprices,
  max_pct_per_order,
  min_dollars_per_order,
  dollars_per_share
)
}
\arguments{
\item{shares_traded}{number of shares traded}

\item{dollars_per_share}{}
}
\description{
US Tiered Minimum Commission Model
}
