% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_commission_futs_backtest.R
\name{wrangle_contracts_on_oi}
\alias{wrangle_contracts_on_oi}
\title{Wrangle Futures Contracts for Simulation using Open Interest}
\usage{
wrangle_contracts_on_oi(contracts)
}
\arguments{
\item{contracts}{A dataframe of futures contract data including columns
ticker, date, close, point_value, open_interest.}
}
\value{
Long dataframe with columns:
symbol: Futures symbol without contract designation (eg ES, GC, ZB, etc)
date: In YYYY-MM-DD format
close_current_contract: Closing USD value of the current contract
(close\emph{point_value)
current_contract: Contract designation for contract with highest open
interest
close_previous_contract: Closing USD value of yesterday's contract with
highest open interest (close}point_value)
previous_contract: Contract designation for contract with highest open
interest yesterday
roll: Boolean specifying whether there was a roll event for the current
observation

Note that unless there was a roll event,
previous_contract == current_contract and
close_previous_contract == close_current_contract
}
\description{
This is a helper function for determining when to roll futures
contracts based on open interest and extracting data required for simulations
that include the roll.

The roll happens the day after a new contract is identified as having the
maximum open interest at the close.

Example: CME gold futures (GC)
2021-01-15: G has highest OI
2021-01-19: J has highest OI (no trading 16-19)
2021-01-20: we roll into J (previous_contract and close_previous_contract
correspond to G, current_contract and close_current_contract to J)
}
\examples{
\dontrun{wrangle_contracts_on_oi(futures)}
}
