% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{positionsFromNoTradeBuffer}
\alias{positionsFromNoTradeBuffer}
\title{Calculate target positions from theoretical weights and trade buffer parameter}
\usage{
positionsFromNoTradeBuffer(
  current_positions,
  current_prices,
  current_theo_weights,
  cap_equity,
  trade_buffer
)
}
\description{
Calculate target positions from theoretical weights and trade buffer parameter
}
